Don't serialize empty sets
If any property of an ISiren* is an IEnumerable, but is created with said IEnumerable being empty, don't bother serializing it. That said, when the object is instantiated, if the value is null then default to an empty IEnumerable. This allows us to still use extension methods on things that may be "null".

For example, if a SirenAction is created with no fields, any calls to theAction.Fields.Whatever will fail, since it's null. This includes things like theAction.Fields.Any(), which one would expect to always work. This default-to-empty-array-but-don't-serialize-empty-array behaviour allows for this, and also prevents serializing empty arrays pointlessly.

Probably worth noting that doing this requires using the "ShouldSerializeFoo" methods, which require fields to implicit, not explicit declarations.