Add ISirenSerializable
Some performance problems came up with using the built-in JsonConvert.SerializeObject due to hitting the (now-quite-expensive) .Equals method for entities. The recommended course of action here is to write our own serializer, which can be significantly faster. In our case, performance on a couple of not-previously-terrible routes improved by a significant margin (20-40%), and one route went from 30+ seconds to <1 second for a response. It's likely the latter case had something else going on too, but hey, across the board performance boosts are nice!