US97770 Validate field type
Siren imposes limits on what a Siren field's `type` can be, but this was not being enforced by this library. As a first step, add SirenFieldType, which includes all the valid field types. This is a non-breaking change, which will be followed up by updating usages of this library to use SirenFieldType instead of a string, followed by a breaking change here to update the constructor to only allow for SirenFieldType.